+++ src/common/src/pgr_assert.cpp
@@ -11,7 +11,7 @@
  * the terms of the MIT License. Please file LICENSE for details.
  *
  ********************************************************************VRP*/
-#include "vrp_assert.h"
+#include "pgr_assert.h"
 #include <exception>
 
 #ifdef assert
