+++ src/ksp/src/KSPDriver.h
@@ -4,7 +4,7 @@
 #include <stdarg.h>
 #include <stdio.h>
 
-#include "pgr_types.h"
+#include "./../../common/src/pgr_types.h"
 
 #ifdef __cplusplus
 extern "C" {
